Ismael - reserva y correccion de producto y relaciones de mapped by medio arreglado

+++ src/main/java/com/example/catpuccino_back/dto/ProductoDTO.java
@@ -1,11 +1,7 @@
 package com.example.catpuccino_back.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Data;
-import lombok.Builder;
-
-
 
 @Data
 @Builder
@@ -13,5 +9,5 @@ public class ProductoDTO {
     private Integer id;
     private String nombre;
     private String descripcion;
-    private Integer precio;
+    private Double precio;
 }
